% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_when.R
\name{dt_case_when}
\alias{dt_case_when}
\title{Case When with data.table}
\usage{
dt_case_when(...)
}
\arguments{
\item{...}{statements of the form: `condition ~ label``, where the label is applied if the condition is met}
}
\description{
Does what `dplyr::case_when()` does, with the same syntax, but with
`data.table::fifelse()` under the hood
}
\examples{

x <- rnorm(1e5)

dt_case_when(
  x < median(x) ~ "low",
  x >= median(x) ~ "high",
  is.na(x) ~ "other"
  )

}
