% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_helpers.R
\name{dt_ends_with}
\alias{dt_ends_with}
\title{Select helpers}
\usage{
dt_ends_with(match)
}
\arguments{
\item{match}{}
}
\value{

}
\description{
These functions allow you to select variables based on their names.
\itemize{
\item \code{dt_starts_with()}: Starts with a prefix
\item \code{dt_starts_with()}: Ends with a suffix
\item \code{dt_contains()}: Contains a literal string
\item \code{dt_everything()}: Matches all variables
}
}
\examples{
library(data.table)
library(tidyfast)

example_dt <- data.table(x = c(1,2,3), y = c(4,5,6), z = c("a", "b", "c"))

example_dt \%>\%
  dt_select(dt_ends_with("z"), dt_everything())
}
