% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{dt_hoist}
\alias{dt_hoist}
\title{Fast Unnesting of Vectors}
\usage{
dt_hoist(dt_, ..., by = NULL)
}
\arguments{
\item{dt_}{the data table to nest}

\item{...}{the columns to unnest (must all be the sample length within ids); use bare names of the variables}

\item{by}{the ID variable to unnest by}
}
\description{
Quickly nest vectors nested in a list column.
}
\examples{

library(data.table)
dt <- data.table(
   x = rnorm(1e5),
   y = runif(1e5),
   nested1 = lapply(1:10, sample, 10, replace = TRUE),
   nested2 = lapply(c("thing1", "thing2"), sample, 10, replace = TRUE),
   id = 1:1e5
   )

dt_hoist(dt,
         nested1, nested2,
         by = id)

}
