% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_variants.R
\name{dt_rename_all}
\alias{dt_rename_all}
\alias{dt_rename_at}
\alias{dt_rename_if}
\title{Rename a selection of variables}
\usage{
dt_rename_all(dt_, .fun, ...)
dt_rename_at(dt_, .vars, .fun, ...)
dt_rename_if(dt_, .predicate, .fun, ...)

dt_rename_at(dt_, .vars, .fun, ...)

dt_rename_if(dt_, .predicate, .fun, ...)
}
\arguments{
\item{dt_}{A data.frame or data.table}

\item{.fun}{Function to pass}

\item{...}{Other arguments for the passed function}

\item{.vars}{\code{list()} of variables for \code{dt_rename_at()} to use}

\item{.predicate}{Predicate to specify columns for \code{dt_rename_if()}}
}
\value{

}
\description{
These scoped variants of `rename()`` operate on a selection of variables
}
