% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{dt_separate}
\alias{dt_separate}
\title{Separate with data.table}
\usage{
dt_separate(dt_, col, into, sep = ".", remove = TRUE, fill = NA,
  fixed = TRUE, by_reference = TRUE, ...)
}
\arguments{
\item{dt_}{the data table (or if not a data.table then it is coerced with as.data.table)}

\item{col}{the column to separate}

\item{into}{the names of the new columns created from splitting `col`.}

\item{sep}{the regular expression stating how `col` should be separated. Default is `.`.}

\item{remove}{should `col` be removed in the returned data table? Default is `TRUE`}

\item{fill}{if empty, fill is inserted. Default is `NA`.}

\item{fixed}{logical. If TRUE match split exactly, otherwise use regular expressions. Has priority over perl.}

\item{by_reference}{should the data table be copied before separating? Default is no (i.e. mutate by reference).}

\item{...}{arguments passed to `data.table::tstrplit()`}
}
\description{
Separates a column of data into others, by splitting based a separator or regular expression
}
\examples{

library(data.table)
d <- data.table(x = c("A.B", "A", "B", "B.A"),
                y = 1:4)
# need to assign when `by_reference = FALSE`
separated <- dt_separate(d, x, c("c1", "c2"), by_reference = FALSE)
separated
# don't need to assign when `by_reference = TRUE` (default)
dt_separate(d, x, c("c1", "c2"), remove = FALSE)
d
dt_separate(d, x, c("c1", "c2"))
d

}
