% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncount.R
\name{dt_uncount}
\alias{dt_uncount}
\title{Uncount}
\usage{
dt_uncount(dt, weights, .remove = TRUE, .id = NULL)
}
\arguments{
\item{dt}{the data table to uncount}

\item{weights}{the counts for each}

\item{.remove}{should the weights variable be removed?}

\item{.id}{an optional new id variable, providing a unique id for each row}
}
\description{
Uncount a counted data table
}
\examples{

library(data.table)

dt_count <- data.table(
  x = LETTERS[1:3],
  w = c(2,1,4)
)
uncount <- dt_uncount(dt_count, w, .id = "id")
uncount   # note that if .remove = TRUE or .id is not null
          # this won't print the first time
          # due to a printing behavior in data.table


}
