% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{dt_unnest}
\alias{dt_unnest}
\title{Fast Unnesting of Data Tables}
\usage{
dt_unnest(dt_, col, by = NULL)
}
\arguments{
\item{dt_}{the data table to nest}

\item{col}{the column to unnest}

\item{by}{the ID variable to unnest by}
}
\description{
Quickly unnest data tables.
}
\examples{

library(data.table)
dt <- data.table(
  x = rnorm(1e5),
  y = runif(1e5),
  grp = sample(1L:3L, 1e5, replace = TRUE)
  )

nested <- dt_nest(dt, grp)
dt_unnest(nested, col = data, by = grp)

}
