% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{dt_unnest_vec}
\alias{dt_unnest_vec}
\title{Fast Unnesting of Vectors}
\usage{
dt_unnest_vec(dt, cols, id, name)
}
\arguments{
\item{dt}{the data table to nest}

\item{cols}{the columns to unnest (must all be the sample length within ids)}

\item{id}{the ID variable to unnest by}

\item{name}{the names of the unnested vectors}
}
\description{
Quickly nest vectors nested in a list column.
}
