% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest.R
\name{nest_dt}
\alias{nest_dt}
\title{Fast Nesting}
\usage{
nest_dt(dt, ..., .key = "data")
}
\arguments{
\item{dt}{the data table to nest}

\item{...}{the variables to group by}

\item{.key}{the name of the list column; default is "data"}
}
\description{
Quickly nest data tables.
}
