% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyfast-package.R
\docType{package}
\name{tidyfast-package}
\alias{tidyfast}
\alias{tidyfast-package}
\title{tidyfast: Fast Tidying of Data}
\description{
Tidying functions built on 'data.table'
    to provide quick and efficient data manipulation with
    minimal overhead.
}
\author{
\strong{Maintainer}: Tyson Barrett \email{t.barrett88@gmail.com} (\href{https://orcid.org/0000-0002-2137-1391}{ORCID})

Other contributors:
\itemize{
  \item Mark Fairbanks [contributor]
  \item Ivan Leung [contributor]
  \item Indrajeet Patil \email{patilindrajeet.science@gmail.com} (\href{https://orcid.org/0000-0003-1995-6531}{ORCID}) (@patilindrajeets) [contributor]
}

}
