% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnest_dt}
\alias{unnest_dt}
\title{Fast Unnesting of Data Tables}
\usage{
unnest_dt(dt, col, id)
}
\arguments{
\item{dt}{the data table to nest}

\item{col}{the column to unnest}

\item{id}{the ID variable to unnest by}
}
\description{
Quickly unnest data tables.
}
