% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{unnest_vec_dt}
\alias{unnest_vec_dt}
\title{Fast Unnesting of Vectors}
\usage{
unnest_vec_dt(dt, cols, id, name)
}
\arguments{
\item{dt}{the data table to nest}

\item{cols}{the columns to unnest}

\item{id}{the ID variable to unnest by}

\item{name}{the names of the unnested vectors}
}
\description{
Quickly nest vectors nested in a list column.
}
