% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest.R
\name{dt_hoist}
\alias{dt_hoist}
\alias{dt_unnest_vec}
\title{Hoist: Fast Unnesting of Vectors}
\usage{
dt_hoist(dt_, ...)
}
\arguments{
\item{dt_}{the data table to unnest}

\item{...}{the columns to unnest (must all be the sample length when unnested); use bare names of the variables}
}
\description{
Quickly unnest vectors nested in list columns. Still experimental (has some potentially unexpected behavior in some situations)!
}
\examples{

library(data.table)
dt <- data.table(
   x = rnorm(1e5),
   y = runif(1e5),
   nested1 = lapply(1:10, sample, 10, replace = TRUE),
   nested2 = lapply(c("thing1", "thing2"), sample, 10, replace = TRUE),
   id = 1:1e5
   )

dt_hoist(dt,
         nested1, nested2,
         by = id)

}
